Reuse existing PWA launch clients
